Working room scanning, xp and leveling

+++ NTU CaptureTheRoom/Views/Login.swift
@@ -59,7 +59,7 @@ struct Login: View {
             }
             let username = data["username"] as! String ?? "unknown"
             let team = data["team"] as? String ?? "unknown"
-            let xp = data["xp"] as! Int
+            let xp = data["xp"] as! CGFloat
             let level = data["level"] as! Int
             let setUpStatus = data["setupstatus"] as? String ?? "notSetUp"
